requirements.txt
aniso8601==7.0.0
appdirs==1.4.4
asgiref==3.4.1
autopep8==1.6.0
blinker==1.4
coverage==6.1.2
Django==3.2.9
django-axes==5.27.0
django-debug-toolbar==3.2.2
django-extra-checks==0.11.0
django-ipware==4.0.0
django-migration-linter==3.0.1
django-split-settings==1.1.0
django-test-migrations==1.1.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
model-bakery==1.3.3
nplusone==1.0.0
promise==2.3
pycodestyle==2.8.0
pytz==2021.3
Rx==1.6.1
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
structlog==21.2.0
text-unidecode==1.3
toml==0.10.2
typing-extensions==3.10.0.2
